requirements.txt
google-adk
google-genai
pydantic
python-dotenv
ollama

#tools/APIs
tavily-python
amadeus
googlemaps
httpx
requests

#frontend
streamlit
streamlit-folium
folium
plotly

#storage
chromadb
sqlalchemy

#utilities
tenacity
